requirements.txt
alabaster==0.7.10
amqp==2.2.2
Babel==2.5.3
certifi==2018.1.18
chardet==3.0.4
diff-match-patch==20121119
Django==1.11
django-admin-view-permission==1.6
django-cors-headers==2.2.0
django-import-export==1.0.0
django-model-utils==3.1.1
django-querysetsequence==0.8
django-redis==4.9.0
django-rest-framework==0.1.0
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
docutils==0.14
gunicorn==19.8.1
MarkupSafe==1.0
odfpy==1.3.6
packaging==17.1
Pygments==2.2.0
PyJWT==1.6.3
pyparsing==2.2.0
PyYAML==3.12
requests==2.18.4
tablib==0.12.1
tornado==5.0.2
urllib3==1.22